Fix review follow-ups for GLM46V fast patch counting test
