+++ src/transformers/models/glm46v/image_processing_glm46v_fast.py
@@ -180,8 +180,6 @@ class Glm46VImageProcessorFast(BaseImageProcessorFast):
         return super().preprocess(images, **kwargs)
 
     def get_number_of_image_patches(self, height: int, width: int, images_kwargs=None):
-        if images_kwargs is None:
-            images_kwargs = {}
         return Glm46VImageProcessor.get_number_of_image_patches(self, height, width, images_kwargs)
